src/streamlit-nl-sql/requirements.txt
# Web Framework
streamlit==1.30.0

# Google Cloud Libraries
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.39.0
db-dtypes==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
altair==5.2.0

# SQL Parsing and Validation
sqlparse==0.4.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
